Use HTML meta to store asset map information
There are a couple downsides to rewriting vendor.js in-place, which
are documented in #36:

- It breaks SRI
- It doesn't update the fingerprint for vendor.js, even though the
  contents have changed.

This patch fixes the issue by loading the asset map information (either
a JSON object or a URL for the asset map) into a HTML meta tag that is
stored in `index.html` instead, then read at runtime by
`getAssetMapData` (the meta tag is formatted as JSON and URL-encoded).

Conceptually, this is fairly similar to the `storeConfigInMeta` option
for the Ember build that stores the app's environment in a meta tag
(https://ember-cli.com/user-guide/#application-configuration).