Passthrough fingerprinted paths
If a path is already fingerprinted (e.g. because the {{asset-map}}
helper was used with a string that got replaced at build time), we can
return the fingerprinted file (i.e. the input we just got), rather than
`undefined`.

Note: this doesn't however require that the fingerprinted path be
correct: if it's not in the asset map, it won't be included.