Show warning for missing `<meta>` tag
When the ember-cli-storybook addon is used there is no meta tag since it is currently generating a different HTML file that does not contain it. While it is okay for ember-cli-ifa to not work correctly in those cases we shouldn't crash the app, but instead show a warning indicating that the addon might not work as expected. Since fingerprinting and ember-cli-ifa are usually only enabled in production builds and ember-cli-storybook is aimed more at development builds that seems like a reasonable tradeoff until ember-cli-storybook can fix how it builds the HTML file.